Make registre page a client component, add redirection if logged in
